+++ source/Tests/QA/ScriptFiles.common.Tests.ps1
@@ -14,11 +14,6 @@ param (
 
 Describe 'Common Tests - Validate Script Files' -Tag 'Script','Common Tests - Validate Script Files' {
 
-    $AddModuleBaseParam = @{
-        NotePropertyName = 'ModuleBase'
-
-    }
-
     $scriptFiles = @(Get-ChildItem $ModuleBase -Recurse -Include *.ps1 | WhereModuleFileNotExcluded).Foreach{
         $_ | Add-Member -NotePropertyName ModuleBase -NotePropertyValue $ModuleBase -PassThru
     }
@@ -37,7 +32,7 @@ Describe 'Common Tests - Validate Script Files' -Tag 'Script','Common Tests - Va
         Write-Debug -Message "... $scriptFile"
         $filePathOutputName = Get-RelativePathFromModuleRoot `
             -FilePath $scriptFile.FullName `
-            -ModuleRootFilePath $_.ModuleBase
+            -ModuleRootFilePath $scriptFile.ModuleBase
 
         Context $filePathOutputName {
             It ('Script file ''{0}'' should not have Byte Order Mark (BOM)' -f $filePathOutputName) {