+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Now skipping isAdmin on non-windows (because creds works differently)
-- Making SpellCheck tests as pending until fixed
+- Moving spellcheck out of the tests
 - Updating ScriptFiles tests to get different module base when from module or from project path
 
 ## [0.4.0] - 2019-12-04